api/requirements.txt
Flask==2.2
requests==2.24.0
Flask_BasicAuth==0.2.0
urllib3==1.25.11
cryptography==3.4.7
SQLAlchemy==1.4.18
flask-sqlalchemy==3.0.2
psycopg2-binary==2.9.6
pydantic==1.10.9
Flask-Migrate==4.0.4
flask_pydantic==0.11.0
Pillow==9.5.0
email-validator==2.0.0
pyjwt==2.7.0
qrcode==7.4.2
pyzbar==0.1.9
flask-cors==3.0.10
ecs-logging==2.0.2
concurrent-log-handler==0.9.24
uWSGI==2.0.21